backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.50.0
python-dotenv>=1.0.0
pydantic>=2.0.0
